refactor change views package name to controller
