+++ src/main/java/gui/controller/AboutUsController.java
@@ -1,4 +1,4 @@
-package gui.views;
+package gui.controller;
 
 
 import gui.Resetable;
