+++ src/main/java/gui/controller/GUI.java
@@ -1,4 +1,4 @@
-package gui.views;
+package gui.controller;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
